### 수정사항 - 1. 메모리 누수 잡음. - 2. KEY 8byte로 변경 - 3. 초기 커밋해놓는 사이즈 증가(추후에 new로 인한 CPU 사용 x)

+++ MMOFIGHTERS/PlayerDefine.h
@@ -1,10 +1,8 @@
 #pragma once
 
-#include <windows.h>
-
 namespace Core
 {
-	typedef UINT64 PLAYER_KEY;
+	typedef unsigned long long PLAYER_KEY;
 	enum class MOVE_DIRECTION
 	{
 		LEFT = 0,
